+++ src/components/bo_avatar/BoAvatar.vue
@@ -4,7 +4,7 @@
 
 <script setup lang="ts">
 import { ref, onBeforeMount, toRefs } from 'vue';
-import type { AvatarComponentProps } from './bo_avatar.type';
+import type { AvatarComponentProps } from './bo_avatar.types';
 import { BoSize } from '@/global';
 
 const props = withDefaults(defineProps<AvatarComponentProps>(), {
