+++ src/components/bo_avatar/bo_avatar.types.ts
