+++ src/components/bo_badge/bo_badge.types.ts
