+++ src/components/bo_icon/index.ts
@@ -1,3 +1,3 @@
 export { default as BoIcon } from './BoIcon.vue';
 export { Icon, IconSizeClasses, icons } from './bo_icon';
-export type { BoIconComponentProps } from './bo_icon.type';
+export type { BoIconComponentProps } from './bo_icon.types';
