+++ src/components/bo_icon_button/index.ts
@@ -5,4 +5,4 @@ export {
 	BoIconButtonSizeClasses,
 	BoIconButtonType,
 } from './bo_icon_button';
-export type { BoIconButtonComponentProps } from './bo_icon_button.type';
+export type { BoIconButtonComponentProps } from './bo_icon_button.types';
