+++ src/components/bo_spinner/bo_spinner.types.ts
