+++ src/components/bo_text/bo_text.types.ts
