+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from '@/components/bo_avatar';
 export * from '@/components/bo_badge';
 export * from '@/components/bo_button';
 export * from '@/components/bo_icon';
